requirements.txt
django==3.0.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-filter==2.3.0
